requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
google-generativeai==0.2.0
Flask-Login==0.6.3
Werkzeug==2.3.6
